Asegurando todos los controladores faltantes

+++ src/main/java/com/itm/biblioteca/repository/MiembroRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MiembroRepository extends JpaRepository<Miembro, Long> {
+public interface MiembroRepository extends JpaRepository<Miembro, Integer> { 
 }
